refactor(crane-mcp): split agent-launch.ts arg parsing from orchestration (531 → 2 modules)
- agent-args.ts (166 LOC): pure arg parsing — resolveAgent, validateAgentBinary,
  stripAgentFlags, extractPassthroughArgs, getStartupPrompt, CRANE_FLAGS constant
- agent-launch.ts (375 LOC): launch orchestrators — launchAgent, launchEngagement,
  assertEngagementScope, spawnAgent, buildChildEnv, applyHermesTranslation + barrel

All callers import from agent-launch.ts unchanged — barrel preserves public API.
